+++ FirebaseCore/Sources/FIRVersion.m
@@ -27,10 +27,6 @@
 #define STR(x) STR_EXPAND(x)
 #define STR_EXPAND(x) #x
 
-@implementation FIRVersion
-
-+ (NSString *)version {
+NSString* FIRFirebaseVersion(void) {
   return [NSString stringWithUTF8String:(const char *const)STR(Firebase_VERSION)];
 }
-
-@end