+++ FirebaseStorage/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 7.0.0
 - [changed] The global variable `FIRStorageVersionString` is deleted.
-  `FirebaseVersion.version()` or `[FIRVersion version]` should be used instead.
+  `FirebaseVersion.version()` or `FIRFirebaseVersion()` should be used instead.
 
 # 3.8.1
 - [fixed] Fixed typo in doc comments (#6485).